fix(aeron): make aeron branch build and tests pass
- sub_threaded: drop redundant `as u64` cast (clippy::unnecessary_cast)
- ChannelSender: notify on drop so receivers can observe sender disconnect
  and surface background-thread exit errors even when the last send has
  already been drained
- ReceiverStream::cycle: drain the channel before checking thread state to
  avoid a race where the thread exits between the check and the drain,
  leaving no one to wake the graph
- rust.yml: install cmake/clang/libclang/uuid/libbsd so --all-features
  clippy can build rusteron-client
- aeron-integration.yml: install protobuf-compiler (transitively required
  via etcd-client in --features aeron-integration-test)

+++ wingfoil/src/adapters/aeron/sub_threaded.rs
@@ -119,7 +119,7 @@ where
                 // Exponential backoff when idle to avoid aggressive spinning.
                 if count == 0 {
                     idle_count = (idle_count + 1).min(20); // Cap at 20 (~1ms sleep)
-                    let micros = (1u64 << idle_count.min(10)) as u64; // 1µs to 1024µs
+                    let micros = 1u64 << idle_count.min(10); // 1µs to 1024µs
                     std::thread::sleep(Duration::from_micros(micros));
                 } else {
                     idle_count = 0;
